backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.2.0
apscheduler==3.10.4
python-dateutil==2.8.2
